requirements.txt
aiohttp==3.4.4
aiohttp-security==0.4.0
aiohttp-session==2.6.0
aiohttp-swagger==1.0.5
asn1crypto==0.24.0
async-timeout==3.0.0
attrs==18.2.0
cffi==1.11.5
chardet==3.0.4
cryptography==2.3.1
easydict==1.8
greenlet==0.4.15
idna==2.7
idna-ssl==1.1.0
Jinja2==2.10
MarkupSafe==1.0
motor==1.2.1
multidict==4.4.2
pycparser==2.19
pymongo==3.6.0
PyYAML==3.13
six==1.11.0
tornado==5.1.1
yarl==1.2.6
